[refactor] duckified clock redux components
