--- client/actions/clock_actions.js
@@ -1,13 +1,0 @@
-'use strict';
-
-const TIMEZONE_ADDED = 'TIMEZONE_ADDED';
-
-const TIMEZONE_CHANGED = 'TIMEZONE_CHANGED';
-
-const TIMEZONE_NOT_SELECTED = 'TIMEZONE_NOT_SELECTED';
-
-module.exports = {
-  TIMEZONE_ADDED,
-  TIMEZONE_CHANGED,
-  TIMEZONE_NOT_SELECTED,
-};
