+++ client/reducers/reducers.js
@@ -5,7 +5,7 @@ const { combineReducers } = require('redux');
 const rootReducer = combineReducers({
   alert: require('../modules/alert').default,
   auth: require('../modules/auth').default,
-  clock: require('./clock_reducer').clockReducer,
+  clock: require('../modules/clock').default,
   dice: require('../modules/dice').default,
   gacha: require('../modules/gacha').default,
   gossip: require('../modules/gossip').default,
